liquidado con rafa

+++ src/main/java/ar/com/playmedia/view/ProductV.java
@@ -1,5 +1,6 @@
 package ar.com.playmedia.view;
 
+import java.text.DecimalFormat;
 import java.util.ArrayList;
 import java.util.Scanner;
 import ar.com.playmedia.controller.ProductC;
@@ -22,7 +23,7 @@ public class ProductV {
         Integer intOption = -1;
         while (intOption != 0) {
             Utilities.clearConsole();
-            System.out.println("Menu de productos");
+            System.out.println("Menu de productos (Hola " + this.oUserM.getName() + ")");
             System.out.println("*************************************\r\n");
             System.out.println("\t1 - Buscar productos");
             System.out.println("\t2 - Agregar nuevo producto");
@@ -33,11 +34,7 @@ public class ProductV {
             System.out.println();
             System.out.println("\t0 - Volver al menu principal");
 
-            try {
-                intOption = Integer.valueOf(oScanner.nextLine());                
-            } catch (Exception e) {
-                intOption = -1;
-            }
+            intOption = Utilities.getNumeric(oScanner, 6);
 
             switch (intOption) {
                 case 1:
@@ -90,7 +87,9 @@ public class ProductV {
         if(iProducts.size() == 0){
             System.out.println("No se encontraron productos");
         }else{
-            System.out.println("ID\t\tNOMBRE\t\tPRECIO\t\tSTOCK\t\tCODIGO");
+            System.out.println("ID    NOMBRE                             PRECIO  CODIGO         STOCK");
+            System.out.println("==    ======                             ======  ======         =====");            
+            System.out.println();
             for(ProductM oProduct : iProducts){
                 System.out.println(oProduct.toString());
             }
@@ -110,9 +109,6 @@ public class ProductV {
         System.out.println("Codigo del producto:");
         oResult.setStrCode(oScanner.nextLine());
 
-        System.out.println("Stock del producto:");
-        String intExistencia = oScanner.nextLine();
-        oResult.setIntStock((intExistencia.equals("")) ? -1 : Integer.valueOf(intExistencia));
 
         if(bitBusqueda){
             System.out.println("Id de producto:");
@@ -120,6 +116,10 @@ public class ProductV {
             oResult.setIntIdProduct((intIdProducto.equals("")) ? 0 : Integer.valueOf(intIdProducto));
         }
         else{
+            System.out.println("Stock del producto:");
+            String intExistencia = oScanner.nextLine();
+            oResult.setIntStock((intExistencia.equals("")) ? -1 : Integer.valueOf(intExistencia));
+
             System.out.println("Precio:");
             oResult.setFloPrice(Float.valueOf(oScanner.nextLine()));
         }
@@ -135,9 +135,22 @@ public class ProductV {
         try {
             ProductC oProductC = new ProductC();
             oProductC.connect();
-            oProductC.saveProduct(oProduct);
-            oProductC.disconnect();
-            System.out.println("Se guardo el registro correctamente");            
+            
+            Utilities.clearConsole();
+            try {
+                ArrayList<ProductM> exists = oProductC.listProduct(new ProductM(0 , "", 0f, oProduct.getStrCode(), 0));
+                if(exists.size() == 0){
+                    oProductC.saveProduct(oProduct);
+                    System.out.println("Se guardo el registro correctamente"); 
+                }
+                else{
+                    System.out.println("El Producto ya existe");
+                }            
+            } catch (Exception e) {
+                System.out.println("Error: " + e.getMessage());
+            }
+            
+            oProductC.disconnect();           
         } catch (Exception e) {            
             System.out.println("Ocurrio un problema al guardar el registro");     
         }
@@ -148,19 +161,50 @@ public class ProductV {
 	private void updateProduct(){
         Utilities.clearConsole();
 
-        System.out.println("Ingrese en id del producto que desea Actualizar");
-        Integer intIdProduct = Integer.valueOf(oScanner.nextLine());
-        ProductM oProduct = new ProductM(intIdProduct, "", 0f, "", -1);
+        ProductM oProduct = new ProductM();
+        ProductC oProductC = new ProductC();
+        ArrayList<ProductM> iProducts = new ArrayList<ProductM>();
+        while(iProducts.size() == 0){
+            Utilities.clearConsole();
+            System.out.println("Ingrese en codigo del producto:");
+            String strCode = oScanner.nextLine();
+            oProduct = new ProductM(0, "", 0f, strCode, -1);
+            oProductC.connect();
+            iProducts =  oProductC.listProduct(oProduct);
+            oProductC.disconnect();
+            if(iProducts.size() == 0){
+                System.out.println("Producto no encontrado");
+                oScanner.nextLine();
+                return;
+            }
+        }
+
+        oProduct = iProducts.get(0);
 
         System.out.println("Que dato quiere actualizar?");
-        System.out.println("1 - Nombre");
-        System.out.println("2 - Precio");
-        System.out.println("3 - Stock");
-        System.out.println("4 - Codigo");
+        System.out.println("0 - Nombre");
+        System.out.println("1 - Precio");
+        System.out.println("2 - Stock");
+        System.out.println("3 - Codigo");
 
-        Integer selection = Integer.valueOf(oScanner.nextLine());
+        Integer selection = Utilities.getNumeric(oScanner, 3);
 
-        System.out.println("Ingrese el valor:");
+        switch (selection) {
+            case 0:                
+                System.out.println("Ingrese el Nombre (" + oProduct.getStrName() + "):");
+                break;                
+            case 1:
+                DecimalFormat decimalFormat = new DecimalFormat("#,##0.00");
+                System.out.println("Ingrese el Precio (" + decimalFormat.format(oProduct.getFloPrice()) + "):");
+                break;            
+            case 2:
+                System.out.println("Ingrese el Stock (" + oProduct.getIntStock() + "):");
+                break;
+            case 3:
+                System.out.println("Ingrese el Codigo (" + oProduct.getStrCode() + "):");
+                break;
+        }
+
         String input = oScanner.nextLine();
 
         oProductC.connect();
@@ -182,17 +226,33 @@ public class ProductV {
                 break;
         }
         oProductC.disconnect();
+        oScanner.nextLine();
 	}
 	
 	private void deleteProduct(){
         Utilities.clearConsole();
-        System.out.println("Ingrese en id del producto que desea eliminar");
+        System.out.println("Ingrese en codigo del producto que desea eliminar");
         Integer intIdProduct = Integer.valueOf(oScanner.nextLine());
         ProductM oProduct = new ProductM(intIdProduct, "", 0f, "", -1);
         try {
             ProductC oProductC = new ProductC();
+
             oProductC.connect();
-            oProductC.deleteProduct(oProduct);
+            System.out.println("Esta seguro de eliminar el producto?");
+            if(Utilities.getYesNo(oScanner)){                
+                try {
+                    ArrayList<ProductM> exists = oProductC.listProduct(new ProductM(0 , "", 0f, oProduct.getStrCode(), 0));
+                    if(exists.size() > 0){
+                        oProductC.deleteProduct(oProduct);
+                        Utilities.clearConsole();
+                    }
+                    else{
+                        System.out.println("El producto ya no se encuentra en existencia");
+                    }            
+                } catch (Exception e) {
+                    System.out.println("Error: " + e.getMessage());
+                }
+            }
             oProductC.disconnect();
             System.out.println("El producto fue eliminado exitosamente");
             oScanner.nextLine();
@@ -213,7 +273,9 @@ public class ProductV {
         if(iProducts.size() == 0){
             System.out.println("No se encontraron productos");
         }else{
-            System.out.println("ID\t\tNOMBRE\t\tPRECIO\t\tSTOCK\t\tCODIGO");
+            System.out.println("ID    NOMBRE                             PRECIO  CODIGO         STOCK");
+            System.out.println("==    ======                             ======  ======         =====");            
+            System.out.println();
             for(ProductM oProduct : iProducts){
                 System.out.println(oProduct.toString());
             }
